requirements.txt
openai
pydantic
pandas
python-dotenv
tabulate
